third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-crash-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: target.createShadowRoot is not a function
PASS unless crash.
